- check that find_lang rev is not outdated
Changed files:
    rpm.spec -> 1.837.2.20

+++ rpm.spec
@@ -25,6 +25,7 @@
 %define	reqdb_ver	4.6.18
 %define	reqpopt_ver	1.10.8
 %define	beecrypt_ver	2:4.1.2-4
+%define	find_lang_rev	1.25
 %define	sover	4.4
 Summary:	RPM Package Manager
 Summary(de.UTF-8):	RPM Packet-Manager
@@ -478,7 +479,7 @@ Requires:	sed
 Requires:	sh-utils
 Requires:	tar
 Requires:	textutils
-Provides:	rpmbuild(find_lang) = 1.25
+Provides:	rpmbuild(find_lang) = %{find_lang_rev}
 Provides:	rpmbuild(monoautodeps)
 Provides:	rpmbuild(noauto) = 3
 %ifarch %{x8664}
@@ -724,6 +725,12 @@ for f in doc{,/ja,/pl}/rpm.8 doc{,/ja,/pl}/rpmbuild.8 ; do
 done
 
 %build
+rev=$(awk '/^#.*Id:.*/{print $4}' scripts/find-lang.sh)
+if [ "$rev" != "%find_lang_rev" ]; then
+	: Update find_lang_rev define to $rev, and retry
+	exit 1
+fi
+
 %if %{with system_libmagic}
 rm -rf file
 %else
